blog-gen/requirements.txt
# Python=3.10
python-dotenv
streamlit
langchain
pypdf
openai
tiktoken
